Refactor argument tokenization to be generic
This changes argument tokenization such that it will parse all flags
with the matching default syntax and any explicitly specified flags.

Additionally, whitespace must now surround the flags for them to be
valid.

Extra convenience methods are added to simplify new operations that
can be useful for commands thanks to this refactor.